Fix that makes reading a pgpass file work when the pgpass file has more than one entry.
The db_getconn function used to open the pgpass file for a second time and uses the attributes (more than just password) from the file instead of what is in the db table. This is after the password is read in from the file. The problem with the code before was that it takes the first entry in the pgpass file that has the correct format. Where as the pervious password related code matches the correct line. So the solution is to simply delete this large block of code as its seems redundant unless we want attributes to override those in the database ( I sure don't want that ).