Fix support for pg 8.1, mostly
This modifies several INSERT/UPDATE ... RETURNING calls to instead use two
queries -- one to do the INSERT or UPDATE, and one to return the currval of a
sequence, which is what RETURNING wanted anyway. Also reworked "DROP TABLE IF
EXISTS" calls, which 8.1 also doesn't like. The tests still won't pass on 8.1,
though, because the logic behind customselect syncs is unsupported in 8.1.
Still working on that one...