Change default reason_file name to end in .txt.
Because it is not a log file, only the most reason reason is written
to it.

But also modify the rason log file to replace the extension from the reason
file with `.log`. So it should no longer be named
`bucardo.restart.reason.log.log` by default.

In other words, the names are:

Reason File: `bucardo.restart.reason.txt`
Rason Log:   `bucardo.restart.reason.log`.

Closes GitHub issue #9.