Test all return values of `ctl()` in postgres tests.
I am still getting test failures, but they are less noisy now that I have it
die after a `sync` fails. The current reason for the failures is that the call
to `goto KID` added in ea4923155f comes before the definition of the KID
label.

In passing, make it clearer that BucardoTesting forces a timeout even when you
pass a `0` to `ctl()`.