Store and use options when restarting after death.
Took me a while to hunt down, but on a server, if I killed the PostgreSQL back
ends, Bucardo eventually died. Once it noticed the back ends were gone, it
tried to restart itself via `exec`. Unfortunately, however, I was using a
start script that passes command-line options, and those options were not
preserved for the `exec` call. It died because it did not have permission to
open the log file, because `--debugdir` was no longer known to it!

So we try to preserve all the command-line options and pass them to
`start_mcp()`. The stripping of nouns from the options is not perfect, but
probably good enough.

We also remove the `dosyncs` stuff from `start_mcp()` and friends, since it
was never used: no parameters are supported for the `start` command, and
nothing was passed to `start_mcp()`. So the feature was never used. Removing
all that stuff, we can just pass in the command-line arguments and use them in
the `exec` call.

Resolves #28.