Fix issues with long table names needing quoting
In bucardo_tablename_maker(), long table/schema names which needed to be
quoted (capitalization or leading digits, say) would cause the wrong MD5
to be calculated, resulting in missing delta/track tables for the tables
in question.

Solve by dequoting the names of the table prior to calculating the MD5
and relying on the later quote_ident() to fix things up.

This only showed up with long table names and ones which needed to be
quoted, so we hadn't encountered this to this point.

Fixes: https://github.com/bucardo/bucardo/issues/125