Fix up postgres version checking regexp
If current working directory not allowed to access, then psql crashes
with error "could not change directory to DIR", and if this DIR contains
numbers, e.g. "DBIx-Safe-1.2.5", then version of postgres will be filled
with this numbers, and command "bucardo install" will be crashed.
The solution of this problem is extending of regexp that used for
finding postgres version.