Remove the herdcheck function and trigger on the herdmap table.
Digging into Github issue #111 showed that the slowdown was caused primarily
by the herdcheck trigger on the herdmap table. Rather than scan the whole table
after each row is added, we do a one-time check right before we commit anytime
we do inserts to the table from the 'bucardo' program.