Add a test showing how bucardo can have a hard time with pgservice.
In an environment in which pgservice must be used, but which also
must have a different service for each system user, bucardo does
not currently work. This is because the pgservice used by the CLI
commands will run as some user, but bucardo pl/Perl functions
will run as the db user (probably postgres). In an environment
where, for example, pgservice defines a certificate auth method,
and requires that every OS user use a distinct service, having
a single pgservice defined in the  bucardo.db table doesn't work.

This test shows that failure.