Allow bucardo add db to take it all in when given a pgservice.
Before, there was an assumption that providing a pgservice on the
command line would mean that every other connection setting on
the command line (host, port, etc) were going to be provided in
the pgservice config. That is incorrect. Fix that assumption.

Also add some unit tests to show this fix, which fails before
the change and succeeds afterwards.